requirements.txt
asgiref==3.8.1
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.0.6
djangorestframework==3.15.1
dlib==19.24.4
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ipdb==0.13.13
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.9.0.80
parso==0.8.4
pillow==10.3.0
prompt_toolkit==3.0.46
pure-eval==0.2.2
Pygments==2.18.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
wheel==0.43.0
